Added VERSION 1.0
1.0:
This version of the game is a fully functional console game. It will greet the player and prompt them
to enter a number that equates to either rock, paper, or scissors. It will then play a round and display
the result to the console. The game will then loop the above sequence 4 more times. After the 5th round
the game will display the overall game winner as well as the game stats and an end game message that
changes depending on the outcome.